1.java8默认垃圾收集器

+++ jvm/src/com/chen/oom/OOMTest.java
@@ -10,7 +10,7 @@ import java.util.Set;
 public class OOMTest {
     public static void main(String[] args) {
         Set<String> set = new HashSet<>();
-        short i = 0;
+        int i = 0;
         while (true) {
             set.add(String.valueOf(i++).intern());
         }
